stm32f1: Added support for mini STLink clone
STLink clones are a good way to get STM32F1-based boards in a great
form-factor and, usually, they make their own SWD pins available via
thru-holes on the board so it's easy to wipe the firmware and then they
can be used for all sorts of things (including as a Black Magic Probe if
you actually bought one to do debugging with).

However, once you have probed the pins with a multimeter they can be
used for all sorts of things!

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>